requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
python-decouple==3.8
requests==2.31.0
pandas==2.1.0